% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure.R
\name{check.and.configure}
\alias{check.and.configure}
\title{Pre Installation Script}
\usage{
check.and.configure(config.file = cargs[length(cargs)], ...)
}
\arguments{
\item{config.file}{The configuration file for the configure script. Defaults to the last parameter given to the script}
}
\description{
Will check the proper setting of Jalgos environment variables and then run `configure`.
}
