% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure.R
\name{check.jalgos.compatible}
\alias{check.jalgos.compatible}
\title{Checks Jalgos Environment}
\usage{
check.jalgos.compatible(
  pkg.name,
  lib.var = "JALGOS_LIB",
  include.var = "JALGOS_INCLUDE",
  config = Sys.getenv()
)
}
\arguments{
\item{lib.var}{Variable that locate jalgos libs}

\item{include.var}{Variable that locate jalgos includes}

\item{config}{Set of variables}
}
\description{
Checks that Jalgos environment variables are set
}
