% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{dependencies}
\alias{dependencies}
\title{Dealing With Dependencies}
\usage{
dependencies(
  libpath = "lib",
  jspackages = list(),
  cran.packages = list(),
  github.packages = list(),
  jslibs = list(),
  force.cran = jsroot.env$reinstall.cran,
  force.github = jsroot.env$reinstall.github,
  force.jalgos = jsroot.env$reinstall.jalgos,
  force.jslibs = jsroot.env$reinstall.jslib,
  force = FALSE,
  ...,
  install.jspackages.fun = install.git
)
}
\arguments{
\item{libpath}{Folder in which}

\item{jspackages}{Jalgos packages. Must have this layout:\cr
list(group1 = list(c("jspack1", \cr
                                 [branch = "aaa"],\cr
                                  ...)),\cr
          group2 = ...)}

\item{cran.packages}{CRAN packages}

\item{...}{To be forwarded to the require.or.install function}
}
\description{
Elegantly deals with package dependencies of a project
}
\seealso{
require.or.install install.git
}
