% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{force.reinstall}
\alias{force.reinstall}
\alias{reinstall.next}
\alias{reinstall.cran.next}
\alias{reinstall.jalgos.next}
\alias{reinstall.jslib.next}
\alias{reinstall.github.next}
\title{Force Reinstall}
\usage{
reinstall.next()

reinstall.cran.next()

reinstall.jalgos.next()

reinstall.jslib.next()

reinstall.github.next()
}
\description{
Force Reinstall
}
\section{Functions}{
\itemize{
\item \code{reinstall.next}: When this function is called all the dependencies package will be reinstall when dependencies is called again

\item \code{reinstall.cran.next}: Will reinstall only the cran packages the next time dependencies is called

\item \code{reinstall.jalgos.next}: Will reinstall only the jalgos packages the next time dependencies is called

\item \code{reinstall.jslib.next}: Will reinstall only the jalgos libraires the next time dependencies is called

\item \code{reinstall.github.next}: Will reinstall only the github packages the next time dependencies is called
}}

