% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start.project.R
\name{init.jsproject}
\alias{init.jsproject}
\title{Initialize project}
\usage{
init.jsproject(data.path, logger.name, template.dir = "project_template/")
}
\arguments{
\item{data.path}{character indicating path to the project data}

\item{logger.name}{character indicating the wanted name for the logger in the Logger Factory}

\item{template.dir}{character indicating the directory name where the template files are located}
}
\description{
Initializes directories and files with data loading, libraries, .gitignore, symbolik link towards the data. To be called from the project repository already created
}
