% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install.git}
\alias{install.git}
\title{Install Git}
\usage{
install.git(
  name,
  group,
  url = make.jalgos.url(name, group, ...),
  install.fun = function(ploc, ...) install.packages(ploc, repos = NULL, ..., type =
    "source"),
  ...
)
}
\arguments{
\item{name}{Name of the package}

\item{group}{Gitlab group to which the package belongs}

\item{url}{Remote url of the package}

\item{install.fun}{Function to use to install the packages}

\item{branch}{specify a branch to checkout}
}
\description{
Installing Jalgos packages with git without using git2r package.
}
