% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install.jalgos.lib}
\alias{install.jalgos.lib}
\title{Install C++ Jalgos Library}
\usage{
install.jalgos.lib(
  name,
  group,
  lib.name = name,
  url = make.jalgos.url(name, group, ...),
  lib.var = "JALGOS_LIB",
  include.var = "JALGOS_INCLUDE",
  force = FALSE,
  ...
)
}
\arguments{
\item{name}{Name of the project}

\item{group}{Group to which the project belong in remote repo}

\item{lib.name}{Name of the library generated by the project}

\item{url}{url of the remote project}

\item{lib.var}{Name of the environment variable containing the directory to which libraries must be installed}

\item{include.var}{Name of the environment variable containing the directory to which include files must be installed}

\item{force}{Force reinstallation even if libraries already exist}
}
\description{
Install the C++ libraries that are required for R packages to run.
}
