% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install.jslib}
\alias{install.jslib}
\title{Installing C++ Jalgos Library}
\usage{
install.jslib(url, pkg.loc = make.tmp.pkg.loc(), ...)
}
\arguments{
\item{url}{url of the package to compile and install}

\item{pkg.loc}{Temporary location where to checkout the package}
}
\description{
Will fetch git project located at url `url` into location in `pkg.loc` and will compile and install the project. Removes `pkg.loc` at the end of the process
}
