% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{reinstall.next}
\alias{reinstall.next}
\title{Force Reinstall}
\usage{
reinstall.next()
}
\description{
When this function is called all the dependencies package will be reinstall when dependencies is called again
}
