% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{dependencies}
\alias{dependencies}
\title{Dealing With Dependencies}
\usage{
dependencies(libpath = "lib", jspackages = list(), cran.packages = list(),
  ...)
}
\arguments{
\item{libpath}{Folder in which}

\item{jspackages}{Jalgos packages. Must have this layout: list(group1 = c("jspack1", ...), group2 = ...)}

\item{cran.packages}{CRAN packages}

\item{...}{To be forwarded to the require.or.install function}
}
\description{
Elegantly deals with package dependencies of a project
}
\seealso{
require.or.install install.git
}
